admin manager

+++ src/main/java/com/example/onlyfanshop_be/dto/BrandDTO.java
@@ -8,6 +8,6 @@ import lombok.*;
 @NoArgsConstructor
 @Builder
 public class BrandDTO {
-    private Long brandID;
+    private Integer brandID;
     private String name;
 }
